muttlib: add mutt_qsort_r
Being able to sort data with reference to auxiliary information passed
in through a local parameter rather than requiring non-local context
through a global variable makes code easier to read.  qsort_r() will
be added in a future version of POSIX [1] with the same signature
present in glibc, and nicely fills this role, but is not portable yet,
so we have to write a shim instead.  Various FreeBSD releases ship
with an incompatible signature for qsort_r (the opaque variable passed
in a different parameter), but have qsort_s with a usable signature
(C11 mentioned qsort_s as optional, and in general the family of *_s
functions hasn't been very widely adopted).  And for platforms that
lack either re-entrant form, as long as our signal handlers and
comparators don't themselves try to call a secondary sort, we don't
care about re-entrancy, and can isolate the use of external data to
static variables in one file rather than global variables across
multiple files.

(Writing a qsort_r replacement on top of qsort where re-entrancy does
matter would require a mutex and shenanigans to support a stack of
nested calls, which gets a lot more complex.)